+++ R/ColumnLinkedMatrix.R
@@ -213,6 +213,10 @@ setMethod("initialize", "ColumnLinkedMatrix", function(.Object, ...) {
     if (length(list) == 0) {
         list[[1]] <- matrix()
     }
+    # Detect non-matrix objects by checking dimensions
+    if (any(unlist(lapply(list, function(x) length(dim(x)) != 2)))) {
+        stop("arguments need to be matrix-like")
+    }
     .Object <- callNextMethod(.Object, list)
     return(.Object)
 })